feat: cid is required in subsystems
We really shouldn't be creating temporary CIDs on the plugin's behalf, anyway. Makes
cleanup a real pain.

BREAKING CHANGE:
Plugins will now need to provide their own CID to PresenceSystem and MessagingSystem. If they
do not have a CID and want to use a temporary CID, they'll need to call `Utils.getCID()` on
their own and use that.